+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 2.9.3
 ----------------
 - Force email fields check using HTML5 input's email type, pattern & placeholder in General_Info.inc.php & StudentsUsersInfo.fnc.php
+- Replace jQuery ScrollToFixed plugin with jQuery FixedMenu in assets/js/jquery-fixedmenu/
 
 Changes in 2.9.2
 ----------------